Optionally redirect other domain names
Put a certificate and key in /etc/authfed like so:
/etc/authfed/example.com-privkey.pem
/etc/authfed/example.com-fullchain.pem

Then map from domains to URLs in targets.edn like so:
{"example.com" "http://www.example.com"}

My test server was previously set up with NGINX as a reverse proxy (ports 80
and 443) in front of Authfed (ports 8080 and 8443). That was okay but now I
want to offer mTLS authentication for the /vault endpoint (for the Terraform
provider) so I need Authfed to bind to port 443 directly.

Apart from being a reverse proxy, NGINX had another job: to serve a few
redirects for some misc domain names I manage. If I can get Authfed to do that
as well then I won't need to spend another $5/month on a t2.micro EC2 instance
just to run NGINX.

Note the use of CAP_NET_BIND_SERVICE to bind to 443 as authfed (i.e. not root).